Moved report class def into its own file per convention; Added ability for trimmed host to output logging to stdout by use of argument passed from command line; Cleaned up a few bugs in calling async routines; Tests pass now.
